+++ MemfaultPackages/bort-ota-lib/src/main/java/com/memfault/bort/ota/lib/Constants.kt
@@ -2,6 +2,8 @@ package com.memfault.bort.ota.lib
 
 const val OTA_PATH = "/data/ota_package/ota.zip"
 const val BORT_SOFTWARE_UPDATE_SETTINGS_PROVIDER = "com.memfault.bort.softwareupdate.settings"
+const val OTA_SETTINGS_COLUMN_LEGACY = 0
+const val OTA_SETTINGS_COLUMN_FULL = 1
 
 const val DEFAULT_STATE_PREFERENCE_FILE = "updater_preferences"
 const val STATE_KEY = "com.memfault.bort.updater.state"