+++ MemfaultPackages/bort-ota-lib/src/main/java/com/memfault/bort/ota/lib/SoftwareUpdateSettingsProvider.kt
@@ -3,8 +3,6 @@ package com.memfault.bort.ota.lib
 import android.content.ContentResolver
 import android.net.Uri
 import com.memfault.bort.shared.SoftwareUpdateSettings
-import kotlinx.serialization.SerializationException
-import kotlinx.serialization.json.Json
 
 interface SoftwareUpdateSettingsProvider {
     fun settings(): SoftwareUpdateSettings?
@@ -22,13 +20,10 @@ class BortSoftwareUpdateSettingsProvider(
             null
         )?.use {
             return if (it.moveToNext()) {
-                val serializedConfig = it.getString(0)
-                try {
-                    Json.decodeFromString(SoftwareUpdateSettings.serializer(), serializedConfig)
-                } catch (ex: SerializationException) {
-                    ex.printStackTrace()
-                    null
-                }
+                // If Bort is recent enough to send new settings fields, use them. See LegacyOtaSettings.
+                val column = if (it.columnCount > OTA_SETTINGS_COLUMN_FULL) OTA_SETTINGS_COLUMN_FULL
+                else OTA_SETTINGS_COLUMN_LEGACY
+                SoftwareUpdateSettings.deserialize(it.getString(column))
             } else {
                 null
             }