+++ MemfaultPackages/bort-ota/src/main/java/com/memfault/bort/ota/AutoInstallRules.kt
@@ -1,0 +1,10 @@
+package com.memfault.bort.ota
+
+/**
+ * Custom rules defining whether an OTA update can be auto-installed right now on this device.
+ */
+fun custom_canAutoInstallOtaUpdateNow(): Boolean {
+    // TODO Customize this method to add and custom logic which determines whether an OTA update can be
+    // auto-installed right now on this device.
+    return true
+}