+++ MemfaultPackages/bort-ota/src/releaseTest/java/com/memfault/bort/ota/TestReceiver.kt
@@ -12,6 +12,7 @@ import kotlinx.coroutines.launch
 
 private const val INTENT_EXTRA_ECHO_STRING = "echo"
 private const val INTENT_EXTRA_BACKGROUND = "background"
+private const val INTENT_EXTRA_MODE = "mode"
 
 /**
  * A broadcast receiver that listens to test events.
@@ -20,6 +21,16 @@ class TestReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
         testLog("TestReceiver action=${intent.action}")
         when (intent.action) {
+            "com.memfault.intent.action.TEST_BORT_OTA_MODE" -> {
+                val modeExtra = intent.getStringExtra(INTENT_EXTRA_MODE)
+                val mode = when (modeExtra) {
+                    "ab" -> TestApp.OtaTestMode.AB
+                    "recovery" -> TestApp.OtaTestMode.Recovery
+                    else -> throw IllegalArgumentException("Invalid test mode")
+                }
+                (context.applicationContext as TestApp).changeMode(mode)
+                testLog("bort-ota updater mode set to $modeExtra")
+            }
             "com.memfault.intent.action.TEST_BORT_OTA_ECHO" -> {
                 testLog("bort-ota echo ${intent.getStringExtra(INTENT_EXTRA_ECHO_STRING)}")
             }