+++ MemfaultPackages/bort/src/main/java/com/memfault/bort/settings/Settings.kt
@@ -105,6 +105,7 @@ interface HttpApiSettings {
     val useMarUpload: Boolean
     val batchMarUploads: Boolean
     val batchedMarUploadPeriod: Duration
+    val useDeviceConfig: Boolean
 }
 
 interface RebootEventsSettings {